Fix LOGX includes and build errors
- Add missing LOGX includes to 7 files that use LOGX macros
- Fix format warnings in debugging_utilities.c (%ld -> %lld for time_t)
- Complete systematic printf/fprintf to LOGX replacement (115 files)
- Fix null pointer crash issues in crash handler and UBUS methods
- Simplify crash handler to prevent recursive crashes
- Add comprehensive error handling and debugging utilities
- All build errors related to missing LOGX includes resolved

+++ src/c/autonomy-daemon/shared/utils/debugging_utilities.c
@@ -155,7 +155,7 @@ void debug_save_state_to_file(const char *filename) {
     }
     
     fprintf(file, "=== DEBUG STATE DUMP ===\n");
-    fprintf(file, "Timestamp: %ld\n", time(NULL));
+    fprintf(file, "Timestamp: %lld\n", (long long)time(NULL));
     fprintf(file, "PID: %d\n", getpid());
     fprintf(file, "Debug Level: %d\n", (int)g_debug_level);
     fprintf(file, "Stack Depth: %d\n", g_debug_stack_depth);
@@ -261,8 +261,8 @@ void debug_emergency_dump(const char *reason) {
     
     // Save to emergency file
     char emergency_file[128];
-    snprintf(emergency_file, sizeof(emergency_file), "/tmp/autonomy_emergency_%d_%ld.log", 
-             getpid(), time(NULL));
+    snprintf(emergency_file, sizeof(emergency_file), "/tmp/autonomy_emergency_%d_%lld.log",
+              getpid(), (long long)time(NULL));
     debug_save_state_to_file(emergency_file);
     
     LOGX_FATAL_MSG("=== END EMERGENCY DUMP ===");
